@@ -4,19 +4,21 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math/rand"
+	"math/rand/v2"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/blokur/harego/v2"
-	"github.com/blokur/harego/v2/mocks"
-	"github.com/blokur/testament"
 	"github.com/google/go-cmp/cmp"
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+
+	"github.com/blokur/testament"
+
+	"github.com/blokur/harego/v2"
+	"github.com/blokur/harego/v2/mocks"
 )
 
 func TestNewConsumer(t *testing.T) {
@@ -30,10 +32,11 @@ func TestNewConsumer(t *testing.T) {
 
 func testNewConsumerBadInput(t *testing.T) {
 	t.Parallel()
+
 	conn := func() (harego.RabbitMQ, error) {
 		return &mocks.RabbitMQ{}, nil
 	}
-	tcs := []struct {
+	testCases := []struct {
 		msg  string
 		conf []harego.ConfigFunc
 	}{
@@ -56,15 +59,14 @@ func testNewConsumerBadInput(t *testing.T) {
 		{"queue name", []harego.ConfigFunc{}},
 	}
 
-	for i, tc := range tcs {
-		tc := tc
-		name := fmt.Sprintf("%d_%s", i, tc.msg)
+	for idx, testCase := range testCases {
+		name := fmt.Sprintf("%d_%s", idx, testCase.msg)
 		t.Run(name, func(t *testing.T) {
 			_, err := harego.NewConsumer(conn,
-				tc.conf...,
+				testCase.conf...,
 			)
 			require.Error(t, err)
-			assert.Contains(t, err.Error(), tc.msg)
+			assert.Contains(t, err.Error(), testCase.msg)
 		})
 	}
 
@@ -92,14 +94,14 @@ func testNewConsumerChannel(t *testing.T) {
 
 func testNewConsumerQos(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	r.On("Channel").Return(ch, nil).Once()
-	ch.On("Qos", mock.Anything, mock.Anything, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil).Once()
+	channel.On("Qos", mock.Anything, mock.Anything, mock.Anything).
 		Return(assert.AnError).Once()
 
-	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomLowerString(10)),
 	)
 	assert.ErrorIs(t, err, assert.AnError)
@@ -107,21 +109,21 @@ func testNewConsumerQos(t *testing.T) {
 
 func testNewConsumerQueueDeclare(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	r.On("Channel").Return(ch, nil)
-	ch.On("Qos", mock.Anything, mock.Anything, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil)
+	channel.On("Qos", mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
-	ch.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
+	channel.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything).
 		Return(amqp.Queue{}, assert.AnError).Once()
 
-	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomString(20)),
 	)
 	assert.ErrorIs(t, err, assert.AnError)
@@ -135,27 +137,27 @@ func testNewConsumerQueueBind(t *testing.T) {
 
 func testNewConsumerQueueBindNoArgs(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	prefetchCount := rand.Intn(9999)
-	prefetchSize := rand.Intn(9999)
+	prefetchCount := rand.IntN(9999)
+	prefetchSize := rand.IntN(9999)
 	queue := testament.RandomString(10)
 
-	r.On("Channel").Return(ch, nil)
-	ch.On("Qos", prefetchCount, prefetchSize, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil)
+	channel.On("Qos", prefetchCount, prefetchSize, mock.Anything).
 		Return(nil).Once()
-	ch.On("QueueDeclare", queue, false, true, mock.Anything,
+	channel.On("QueueDeclare", queue, false, true, mock.Anything,
 		true, mock.Anything).
 		Return(amqp.Queue{}, nil).Once()
-	ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
-	ch.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
+	channel.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
+	channel.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
 		Return(assert.AnError).Once()
 
-	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.PrefetchCount(prefetchCount),
 		harego.PrefetchSize(prefetchSize),
 		harego.QueueName(queue),
@@ -168,36 +170,37 @@ func testNewConsumerQueueBindNoArgs(t *testing.T) {
 
 func testNewConsumerQueueBindArgs(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	prefetchCount := rand.Intn(9999)
-	prefetchSize := rand.Intn(9999)
+	prefetchCount := rand.IntN(9999)
+	prefetchSize := rand.IntN(9999)
 	queue := testament.RandomString(10)
 	args := map[string]interface{}{
 		"arg1": "val1",
 		"arg2": "val2",
 	}
 
-	r.On("Channel").Return(ch, nil)
-	ch.On("Qos", prefetchCount, prefetchSize, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil)
+	channel.On("Qos", prefetchCount, prefetchSize, mock.Anything).
 		Return(nil).Once()
-	ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
-	ch.On("QueueDeclare", queue, false, true, mock.Anything,
+	channel.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
+	channel.On("QueueDeclare", queue, false, true, mock.Anything,
 		true, mock.MatchedBy(func(a amqp.Table) bool {
 			if diff := cmp.Diff(amqp.Table(args), a); diff != "" {
 				t.Errorf("(-want +got):\n%s", diff)
 			}
+
 			return true
 		})).
 		Return(amqp.Queue{}, nil).Once()
-	ch.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
+	channel.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
 		Return(assert.AnError).Once()
 
-	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	_, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.PrefetchCount(prefetchCount),
 		harego.PrefetchSize(prefetchSize),
 		harego.QueueName(queue),
@@ -226,32 +229,32 @@ func testClientConsume(t *testing.T) {
 
 func testClientConsumeChannelError(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	r.On("Channel").Return(ch, nil)
-	ch.On("Qos", mock.Anything, mock.Anything, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil)
+	channel.On("Qos", mock.Anything, mock.Anything, mock.Anything).
 		Return(nil)
-	ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything).
 		Return(amqp.Queue{}, nil).Once()
-	ch.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
+	channel.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
+	channel.On("NotifyClose", mock.Anything).Return(make(chan *amqp.Error, 10))
 
 	queueName := testament.RandomString(10)
 	consumerName := testament.RandomString(10)
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(queueName),
 		harego.ConsumerName(consumerName),
 		harego.NoWait,
 	)
 	require.NoError(t, err)
 
-	ch.On("Consume",
+	channel.On("Consume",
 		queueName,
 		consumerName,
 		mock.Anything, mock.Anything, mock.Anything,
@@ -259,25 +262,25 @@ func testClientConsumeChannelError(t *testing.T) {
 		mock.Anything,
 	).Return(nil, assert.AnError).Once()
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
 
-	err = cons.Consume(ctx, func(*amqp.Delivery) (a harego.AckType, delay time.Duration) { return 0, 0 })
+	err = cons.Consume(ctx, func(*amqp.Delivery) (harego.AckType, time.Duration) { return harego.AckTypeAck, 0 })
 	assert.ErrorIs(t, err, assert.AnError)
 }
 
 func testClientConsumeNilHandler(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := getPassingChannel(t, 1)
-	r.On("Channel").Return(ch, nil)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := getPassingChannel(t, 1)
+	rabbitCli.On("Channel").Return(channel, nil)
 
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomString(10)),
 	)
 	require.NoError(t, err)
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
 
 	err = cons.Consume(ctx, nil)
@@ -286,68 +289,74 @@ func testClientConsumeNilHandler(t *testing.T) {
 
 func testClientConsumeCancelledContext(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := mocks.NewChannel(t)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := mocks.NewChannel(t)
 
-	r.On("Channel").Return(ch, nil)
-	ch.On("Qos", mock.Anything, mock.Anything, mock.Anything).
+	rabbitCli.On("Channel").Return(channel, nil)
+	channel.On("Qos", mock.Anything, mock.Anything, mock.Anything).
 		Return(nil)
-	ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything).
 		Return(nil)
-	ch.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+	channel.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything).
 		Return(amqp.Queue{}, nil).Once()
-	ch.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
+	channel.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
 		Return(nil).Once()
-	ch.On("NotifyClose", mock.Anything).
+	channel.On("NotifyClose", mock.Anything).
 		Return(make(chan *amqp.Error, 10))
 
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomString(10)),
 	)
 	require.NoError(t, err)
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
+
 	delivery := make(chan amqp.Delivery, 1)
 	delivery <- amqp.Delivery{Body: []byte("first message")}
 
-	ch.On("Consume", mock.Anything, mock.Anything, mock.Anything,
+	channel.On("Consume", mock.Anything, mock.Anything, mock.Anything,
 		mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 	).Return((<-chan amqp.Delivery)(delivery), nil).Once()
 
-	var wg sync.WaitGroup
-	wg.Add(1)
+	var wGroup sync.WaitGroup
 
-	err = cons.Consume(ctx, func(d *amqp.Delivery) (a harego.AckType, delay time.Duration) {
+	wGroup.Add(1)
+
+	err = cons.Consume(ctx, func(d *amqp.Delivery) (harego.AckType, time.Duration) {
 		select {
 		case <-ctx.Done():
 			t.Errorf("didn't expect to receive a call for: %q", d.Body)
-			return 0, 0
+			return harego.AckTypeAck, 0
 		default:
 		}
+
 		cancel()
+
 		delivery <- amqp.Delivery{Body: []byte("second message")}
-		wg.Done()
-		return 0, 0
+
+		wGroup.Done()
+
+		return harego.AckTypeAck, 0
 	})
 	require.ErrorIs(t, err, context.Canceled)
 
 	assert.Eventually(t, func() bool {
-		wg.Wait()
+		wGroup.Wait()
 		return true
 	}, 10*time.Second, 10*time.Millisecond)
 }
 
 func testClientConsumeAlreadyClosed(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := getPassingChannel(t, 1)
-	r.On("Channel").Return(ch, nil)
-	r.On("Close").Return(nil)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := getPassingChannel(t, 1)
+	rabbitCli.On("Channel").Return(channel, nil)
+	rabbitCli.On("Close").Return(nil)
 
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomString(10)),
 	)
 	require.NoError(t, err)
@@ -355,11 +364,12 @@ func testClientConsumeAlreadyClosed(t *testing.T) {
 	err = cons.Close()
 	require.NoError(t, err)
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
-	err = cons.Consume(ctx, func(d *amqp.Delivery) (a harego.AckType, delay time.Duration) {
+
+	err = cons.Consume(ctx, func(d *amqp.Delivery) (harego.AckType, time.Duration) {
 		t.Errorf("didn't expect to receive a call for: %q", d.Body)
-		return 0, 0
+		return harego.AckTypeAck, 0
 	})
 	assert.ErrorIs(t, err, harego.ErrClosed)
 }
@@ -372,12 +382,12 @@ func testClientConsumeHandlePanics(t *testing.T) {
 
 func testClientConsumeHandlePanicsDefaultLogger(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := getPassingChannel(t, 1)
-	r.On("Channel").Return(ch, nil)
-	r.On("Close").Return(nil)
+	rabbitcli := mocks.NewRabbitMQ(t)
+	channel := getPassingChannel(t, 1)
+	rabbitcli.On("Channel").Return(channel, nil)
+	rabbitcli.On("Close").Return(nil)
 
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitcli, nil },
 		harego.QueueName(testament.RandomString(10)),
 	)
 	require.NoError(t, err)
@@ -387,10 +397,12 @@ func testClientConsumeHandlePanicsDefaultLogger(t *testing.T) {
 		assert.NoError(t, err)
 	}()
 
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	ctx, cancel := context.WithTimeout(t.Context(), time.Second)
 	defer cancel()
-	err = cons.Consume(ctx, func(*amqp.Delivery) (a harego.AckType, delay time.Duration) {
+
+	err = cons.Consume(ctx, func(*amqp.Delivery) (harego.AckType, time.Duration) {
 		defer cancel()
+
 		panic(assert.AnError)
 	})
 	assert.ErrorIs(t, err, context.Canceled)
@@ -398,13 +410,13 @@ func testClientConsumeHandlePanicsDefaultLogger(t *testing.T) {
 
 func testClientConsumeHandlePanicsWithLogger(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := getPassingChannel(t, 1)
-	r.On("Channel").Return(ch, nil)
-	r.On("Close").Return(nil)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := getPassingChannel(t, 1)
+	rabbitCli.On("Channel").Return(channel, nil)
+	rabbitCli.On("Close").Return(nil)
 
 	logger := newMockLogger()
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomString(10)),
 		harego.Logger(logger.logger),
 	)
@@ -415,10 +427,12 @@ func testClientConsumeHandlePanicsWithLogger(t *testing.T) {
 		assert.NoError(t, err)
 	}()
 
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+	ctx, cancel := context.WithTimeout(t.Context(), 2*time.Second)
 	defer cancel()
-	cons.Consume(ctx, func(*amqp.Delivery) (a harego.AckType, delay time.Duration) {
+
+	cons.Consume(ctx, func(*amqp.Delivery) (harego.AckType, time.Duration) {
 		defer cancel()
+
 		panic(assert.AnError)
 	})
 
@@ -434,12 +448,12 @@ func testClientClose(t *testing.T) {
 
 func testClientCloseAlreadyClosed(t *testing.T) {
 	t.Parallel()
-	r := mocks.NewRabbitMQ(t)
-	ch := getPassingChannel(t, 1)
-	r.On("Channel").Return(ch, nil)
-	r.On("Close").Return(nil)
+	rabbitCli := mocks.NewRabbitMQ(t)
+	channel := getPassingChannel(t, 1)
+	rabbitCli.On("Channel").Return(channel, nil)
+	rabbitCli.On("Close").Return(nil)
 
-	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 		harego.QueueName(testament.RandomLowerString(10)),
 	)
 	require.NoError(t, err)
@@ -453,10 +467,11 @@ func testClientCloseAlreadyClosed(t *testing.T) {
 
 func testClientCloseErrors(t *testing.T) {
 	t.Parallel()
+
 	err1 := errors.New(testament.RandomString(10))
 	err2 := errors.New(testament.RandomString(10))
 
-	tcs := map[string]struct {
+	testCases := map[string]struct {
 		channelErr error
 		connErr    error
 		wantErrs   []error
@@ -467,37 +482,37 @@ func testClientCloseErrors(t *testing.T) {
 		"both errors": {err1, err2, []error{err1, err2}},
 	}
 
-	for name, tc := range tcs {
-		tc := tc
+	for name, testCase := range testCases {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			r := mocks.NewRabbitMQ(t)
-			ch := mocks.NewChannel(t)
+			rabbitCli := mocks.NewRabbitMQ(t)
+			channel := mocks.NewChannel(t)
 
-			r.On("Channel").Return(ch, nil)
-			ch.On("Qos", mock.Anything, mock.Anything, mock.Anything).
+			rabbitCli.On("Channel").Return(channel, nil)
+			channel.On("Qos", mock.Anything, mock.Anything, mock.Anything).
 				Return(nil)
-			ch.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+			channel.On("ExchangeDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 				mock.Anything, mock.Anything, mock.Anything).
 				Return(nil).Once()
-			ch.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
+			channel.On("QueueDeclare", mock.Anything, mock.Anything, mock.Anything, mock.Anything,
 				mock.Anything, mock.Anything).
 				Return(amqp.Queue{}, nil).Once()
-			ch.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
+			channel.On("QueueBind", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).
 				Return(nil).Once()
-			ch.On("NotifyClose", mock.Anything).
+			channel.On("NotifyClose", mock.Anything).
 				Return(make(chan *amqp.Error, 10))
 
-			cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return r, nil },
+			cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) { return rabbitCli, nil },
 				harego.QueueName(testament.RandomLowerString(10)),
 			)
 			require.NoError(t, err)
 
-			ch.On("Close").Return(tc.channelErr)
-			r.On("Close").Return(tc.connErr)
+			channel.On("Close").Return(testCase.channelErr)
+			rabbitCli.On("Close").Return(testCase.connErr)
+
 			err = cons.Close()
 
-			for _, e := range tc.wantErrs {
+			for _, e := range testCase.wantErrs {
 				assert.ErrorIsf(t, err, e, "want %v, got %v", e, err)
 			}
 		})
@@ -506,6 +521,7 @@ func testClientCloseErrors(t *testing.T) {
 
 func testClientCloseMultipleTimes(t *testing.T) {
 	t.Parallel()
+
 	cons, err := harego.NewConsumer(func() (harego.RabbitMQ, error) {
 		return &mocks.RabbitMQSimple{}, nil
 	}, harego.QueueName(testament.RandomLowerString(10)))
@@ -513,7 +529,8 @@ func testClientCloseMultipleTimes(t *testing.T) {
 
 	err = cons.Close()
 	require.NoError(t, err)
-	for i := 0; i < 100; i++ {
+
+	for range 100 {
 		err = cons.Close()
 		assert.ErrorIs(t, err, harego.ErrClosed)
 	}
