@@ -1,5 +1,3 @@
-//go:build integration
-
 package harego_test
 
 import (
@@ -11,17 +9,24 @@ import (
 	"testing"
 	"time"
 
-	"github.com/blokur/harego/v2"
-	"github.com/blokur/testament"
 	"github.com/google/go-cmp/cmp"
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/testcontainers/testcontainers-go"
+
+	"github.com/blokur/testament"
+
+	"github.com/blokur/harego/v2"
 )
 
 func TestIntegPublisher(t *testing.T) {
 	t.Parallel()
+
+	if testing.Short() {
+		t.Skip()
+	}
+
 	t.Run("Publish", testIntegPublisherPublish)
 	t.Run("PublishWorkers", testIntegPublisherPublishWorkers)
 	t.Run("Reconnect", testIntegPublisherReconnect)
@@ -30,7 +35,8 @@ func TestIntegPublisher(t *testing.T) {
 
 func testIntegPublisherPublish(t *testing.T) {
 	t.Parallel()
-	tcs := []struct {
+
+	testCases := []struct {
 		total   int
 		workers int
 	}{
@@ -44,18 +50,19 @@ func testIntegPublisherPublish(t *testing.T) {
 		{10, 100},
 		{20, 1000},
 	}
-	for _, tc := range tcs {
-		tc := tc
-		name := fmt.Sprintf("%dWorkers/%dMessages/", tc.workers, tc.total)
+
+	for _, testCase := range testCases {
+		name := fmt.Sprintf("%dWorkers/%dMessages/", testCase.workers, testCase.total)
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			testIntegPublisherPublishConcurrent(t, tc.total, tc.workers)
+			testIntegPublisherPublishConcurrent(t, testCase.total, testCase.workers)
 		})
 	}
 }
 
 func testIntegPublisherPublishConcurrent(t *testing.T, total, workers int) {
 	t.Helper()
+
 	exchange := "test." + testament.RandomString(20)
 	queueName := "test." + testament.RandomString(20)
 	routingKey := "test." + testament.RandomString(20)
@@ -64,24 +71,33 @@ func testIntegPublisherPublishConcurrent(t *testing.T, total, workers int) {
 		harego.Workers(workers),
 		harego.RoutingKey(routingKey),
 	)
-	var wg sync.WaitGroup
-	wg.Add(total)
-	for i := 0; i < total; i++ {
+
+	var wGroup sync.WaitGroup
+
+	wGroup.Add(total)
+
+	for range total {
 		go func() {
-			defer wg.Done()
+			defer wGroup.Done()
+
 			err := pub.Publish(&amqp.Publishing{
 				Body: []byte(randomBody(1)),
 			})
 			assert.NoError(t, err)
 		}()
 	}
-	wg.Wait()
+
+	wGroup.Wait()
 }
 
+//nolint:dupl // This is not a duplicate of the CONSUMER.
 func testIntegPublisherPublishWorkers(t *testing.T) {
+	t.Parallel()
+
 	if testing.Short() {
 		t.Skip("slow test")
 	}
+
 	var (
 		exchange1 = "test." + testament.RandomString(20)
 		exchange2 = "test." + testament.RandomString(20)
@@ -92,13 +108,14 @@ func testIntegPublisherPublishWorkers(t *testing.T) {
 		cons      *harego.Consumer
 	)
 
-	err = r.Do(func() error {
+	err = retryConfig.Do(func() error {
 		_, addr = getContainer(t)
 		return nil
 	}, func() error {
 		_, pub, err = getConsumerPublisherWithAddr(t, addr, exchange1, queueName,
 			harego.Workers(10),
 		)
+
 		return err
 	}, func() error {
 		cons, _, err = getConsumerPublisherWithAddr(t, addr, exchange2, queueName)
@@ -106,13 +123,13 @@ func testIntegPublisherPublishWorkers(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	var (
-		want []string
-		mu   sync.RWMutex
-		got  []string
-	)
+	var mutex sync.RWMutex
+
 	total := 1000
-	for i := 0; i < total; i++ {
+	want := make([]string, 0, total)
+	got := make([]string, 0, total)
+
+	for i := range total {
 		msg := fmt.Sprintf("Message: [i:%d]", i)
 		want = append(want, msg)
 		err := pub.Publish(&amqp.Publishing{
@@ -121,40 +138,52 @@ func testIntegPublisherPublishWorkers(t *testing.T) {
 		require.NoError(t, err)
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
+
 	go func() {
 		err := cons.Consume(ctx, func(msg *amqp.Delivery) (harego.AckType, time.Duration) {
-			mu.Lock()
-			defer mu.Unlock()
+			mutex.Lock()
+			defer mutex.Unlock()
+
 			got = append(got, string(msg.Body))
+
 			return harego.AckTypeAck, 0
 		})
 		assert.ErrorIs(t, err, context.Canceled)
 	}()
+
 	assert.Eventually(t, func() bool {
-		mu.RLock()
-		defer mu.RUnlock()
+		mutex.RLock()
+		defer mutex.RUnlock()
+
 		if len(want) == len(got) {
 			cancel()
 			return true
 		}
+
 		return false
 	}, time.Minute, 10*time.Millisecond)
-	mu.RLock()
-	defer mu.RUnlock()
+
+	mutex.RLock()
+	defer mutex.RUnlock()
+
 	if diff := cmp.Diff(want, got, testament.StringSliceComparer); diff != "" {
 		t.Errorf("(-want +got):\\n%s", diff)
 	}
 }
 
+//nolint:funlen // This is an integrateion test.
 func testIntegPublisherReconnect(t *testing.T) {
+	t.Parallel()
+
 	if testing.Short() {
 		t.Skip("slow test")
 	}
+
 	var (
 		total     = 200
-		wg        sync.WaitGroup
+		wGroup    sync.WaitGroup
 		exchange  = "test." + testament.RandomString(20)
 		queueName = "test." + testament.RandomString(20)
 		container testcontainers.Container
@@ -164,7 +193,7 @@ func testIntegPublisherReconnect(t *testing.T) {
 		err       error
 	)
 
-	err = r.Do(func() error {
+	err = retryConfig.Do(func() error {
 		container, addr = getContainer(t)
 		return nil
 	}, func() error {
@@ -173,17 +202,21 @@ func testIntegPublisherReconnect(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
 
 	// we want the goroutine to wait for us before publishing.
 	restart := make(chan chan struct{})
-	wg.Add(1)
+
+	wGroup.Add(1)
+
 	go func() {
-		defer wg.Done()
+		defer wGroup.Done()
+
 		for {
 			select {
 			case ch := <-restart:
+				//nolint:contextcheck // See func documentation
 				restartRabbitMQ(t, container)
 				close(ch)
 			case <-ctx.Done():
@@ -191,18 +224,24 @@ func testIntegPublisherReconnect(t *testing.T) {
 			}
 		}
 	}()
-	wg.Add(1)
+
+	wGroup.Add(1)
+
 	var calls int32
+
 	go func() {
-		defer wg.Done()
+		defer wGroup.Done()
+
 		assert.NotPanics(t, func() {
 			i := 0
 			err := cons.Consume(ctx, func(*amqp.Delivery) (harego.AckType, time.Duration) {
 				i++
-				g := atomic.AddInt32(&calls, 1)
-				if g >= int32(total) {
+
+				g := int(atomic.AddInt32(&calls, 1))
+				if g >= total {
 					cancel()
 				}
+
 				return harego.AckTypeAck, 0
 			})
 			assert.ErrorIs(t, err, context.Canceled)
@@ -211,34 +250,41 @@ func testIntegPublisherReconnect(t *testing.T) {
 
 	var restarted bool
 
-	for i := 0; i < total; i++ {
+	for pubID := 0; pubID < total; pubID++ {
 		if restarted {
 			time.Sleep(time.Second / 5)
+
 			pub, err = harego.NewPublisher(harego.URLConnector(addr),
 				harego.ExchangeName(exchange),
 			)
 			if err == nil {
 				restarted = false
 			}
-			i--
+
+			pubID--
+
 			continue
 		}
-		err := pub.Publish(&amqp.Publishing{
+
+		err = pub.Publish(&amqp.Publishing{
 			Body: []byte(testament.RandomString(10)),
 		})
 		if err != nil {
-			i--
+			pubID--
 		}
-		if i > 0 && i%(total/5) == 0 {
+
+		if pubID > 0 && pubID%(total/5) == 0 {
 			restarted = true
+
 			msg := make(chan struct{})
 			restart <- msg
+
 			<-msg
 		}
 	}
 
 	assert.Eventually(t, func() bool {
-		wg.Wait()
+		wGroup.Wait()
 		return true
 	}, 2*time.Minute, 50*time.Millisecond)
 
@@ -247,42 +293,50 @@ func testIntegPublisherReconnect(t *testing.T) {
 
 func testIntegPublisherClose(t *testing.T) {
 	t.Parallel()
+
 	exchange := "test." + testament.RandomString(20)
 	queueName := "test." + testament.RandomString(20)
 
 	total := 10
+
 	_, pub := getConsumerPublisher(t, exchange, queueName, harego.Workers(total))
 	defer pub.Close()
 
-	var wg sync.WaitGroup
-	for i := 0; i < total-1; i++ {
-		wg.Add(1)
+	var wGroup sync.WaitGroup
+	for range total - 1 {
+		wGroup.Add(1)
+
 		go func() {
-			defer wg.Done()
+			defer wGroup.Done()
+
 			err := pub.Publish(&amqp.Publishing{
 				Body: []byte(testament.RandomString(20)),
 			})
 			if err == nil {
 				return
 			}
+
 			assert.True(t, errors.Is(err, harego.ErrClosed) || errors.Is(err, context.Canceled))
 		}()
 	}
 
-	wg.Add(total / 2)
-	for i := 0; i < total/2; i++ {
+	wGroup.Add(total / 2)
+
+	for range total / 2 {
 		go func() {
-			defer wg.Done()
+			defer wGroup.Done()
+
 			err := pub.Close()
 			if err == nil {
 				return
 			}
+
 			assert.True(t, errors.Is(err, harego.ErrClosed) || errors.Is(err, context.Canceled))
 		}()
 	}
 
 	assert.Eventually(t, func() bool {
-		wg.Wait()
+		wGroup.Wait()
 		return true
 	}, 2*time.Minute, 30*time.Millisecond)
 }
